Added process_input and utils
