+++ Makefile
@@ -6,13 +6,16 @@
 #    By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/09/03 22:33:20 by irychkov          #+#    #+#              #
-#    Updated: 2024/10/23 22:34:12 by irychkov         ###   ########.fr        #
+#    Updated: 2024/10/24 16:32:07 by irychkov         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
 NAME = minishell
 
-SRCS = src/main.c src/split_args.c src/split_args_helper.c \
+SRCS = src/main.c \
+		src/process_input.c src/process_input_utilities.c \
+		src/process_input_utilities2.c \
+		src/split_args.c src/split_args_helper.c \
 		src/init_pipes.c src/init_cmd.c src/init_shell.c \
 		src/here_doc.c src/here_doc2.c \
 		src/parse_add_spaces.c \
